# 005-job-portal/app/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Market-Place",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get All Jobs",
        "description": "Get All Jobs",
        "operationId": "getAllJobs",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create a Job",
        "description": "Create a Job",
        "operationId": "createAJob",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Single Job",
        "description": "Get Single Job",
        "operationId": "getSingleJob",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Jobs"
        ],
        "summary": "Update a Job",
        "description": "Update a Job",
        "operationId": "updateAJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "company": {
                    "type": "string",
                    "example": ""
                  }
                }
              },
              "examples": {
                "Update a Job": {
                  "value": {
                    "company": ""
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete a Job",
        "description": "Delete a Job",
        "operationId": "deleteAJob",
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "Bearer"
      }
    }
  },
...
}